drop write macro in favor of IO objects + fmt::Write
closes #4
closes #5
closes #6
closes #7
closes #11
